Fixes the problem with the kernel not finding the setting
